Fixed issue where exceptions caused agent to hang.
This was due to the way that the mailbox processor "handles"
exceptions, i.e. it swallows them. Since the agent runs in its own
thread, attempting to reraise the exception would have no affect on the
caller thread.

The solution was to define a discriminated union with a success and
failure case. The exception is then caught in the async workflow and
sent across the reply channel, where it is then reraised. The caller
thread is able to handle the exception now because the async workflow
was converted to a task (and back).

Refer to:
  http://stackoverflow.com/questions/10805035
  http://stackoverflow.com/questions/10881464