Added database and collection fields to scope.
Obeys the assumption that the user will only interact with the Scope
record type via the provided functions in the Query module. It is
possible to inspect the contents of the record type and change the
operatio, e.g apply to a different collection.

Sure would be nice to have compiler flag akin to -opaque as we have
with Erlang.

Integrated previous work done for the Fluent API.