Changed format of expression with code quotations.
Now defined as a lambda expression that takes a single parameter and
returns a boolean. Still need to add pattern matching clauses to handle
user-defined types/classes, but style is relatively the same. May also
be able to take advantage of LINQ support in existing .NET driver.

    <@ fun x -> x?name = "John" && x?age > 20 @> |> bson ->
        { "$and" : [{ "name" : "John" }, { "age" : { "$gt" : 20 } }] }

Added support for dynamic getter on BsonDocument instances, which
allows the user to compare a BsonValue to any type. Note that the
parsing of the quotation will automatically call BsonValue.Create on
the value.