Added support for $inc operator in quotations.
Fixed issue where was not pattern matching to return the value of the
dynamic assignment operator.

Defined function that transforms a single update operation, i.e. an
element from the unit list, into the corresponding BsonElement. The
mapping could be done in parallel, but probably not beneficial in terms
of performance.

Need to do manipulation of the BsonElements via fold to combine the
documents of elements with the same keys, or else the BsonDocument
constructor with raise an exception.