Defined a convention for record and option types.
Uses BsonClassMap for serialization, and maps the constructor as well
as the members. The convention pack is not registered with any filter
as each case (record type, option type, discriminated union) will do
the necessary checks in the `Apply` member.

Moved F#-specific BSON extensions into separate assembly. Although they
should be usable from C# as long as FSharp.Core has been added as a
reference to the project.