Extended support for typed collections and update operations.
Note that in F# 3.0, the type inference system requires an annotation
on the document when used by other members in the computation
expression, regardless of that fact it maintains the variable space.
This appears to have been resolved in F# 3.1 (see the Visual Studio F#
Team Blog).

The return type of the MongoQueryBuilder will be changed to
MongoOperationResult, which will either return a Scope type or actually
execute the operation specified (depending on query/update/etc.).