Added support for $unset operator in mongo expression.
Needs to create a code quotation of `NewUnionCase (None)`. Uses
reflection on the option type to get the union case.

Currently does not enforce that the field resolves to an option type,
as is done with the typed code quotations, but easily could by changing
the type signature to 'a -> 'b option.