Extended support for transforming update operations.
Changed design, so rather than convert to the format recognized by the
existing code quotation parsers, simply constuct the BsonElements here.
This was done due to some complications with the type validation that
goes on with the variables and expressions, especially with the `|>`
(forward-pipe) operator.

Instead, there is now a ``fake'' parser that you can partially apply
the resulting BsonElement to, and still accepts the var, field, and
expression parameters for consistency with the existing structure.