Added support for $each and $slice operators in mongo expression.
Defined two separate custom operators: `pushEach` and `addToSetEach`
for combination with $push and $addToSet respectively.

In order to enforce that the `slice` custom operator is only used
(immediately) after a `xxxEach` operator, a new interface that extends
from IMongoUpdatable was defined.

Note that similar to the other operations in the Fluent Spec, the last
assigned value is the one that is used, e.g. with

    mongo {
        for x in clctn do
        update
        pushEach x?tags [ "appliances"; "school"; "book" ]
        slice -4
        slice -2
    }

the `tags` field will contain only the last **two** elements.