Updated version of .NET Core driver; resolves #17.
Now points to mongodb/mongo-csharp-driver@60f193f.

Also changed backbone and operation settings to option types.

This way can use the defaults that are set in the .NET Core driver,
rather than duplicating the values and trying to keep both versions in
sync.

Only assigns the value to the property if it is a Some case.

Added configuration options to MongoClient. Can specify a list of
hostnames, and also a replica set name.