Refactored $sort implementation for code quotations.
Instead of using an intentionally unchecked query to compare a field to
the value +/- 1 to follow the semantics of the shell operator, defined
new functions that match the conventions of the computation expression
  - `Update.sortListBy`
  - `Update.sortListByDescending`
  - `Update.thenListBy`
  - `Update.thenListByDescending`

Note that the query/update parser functions are no longer dependent on
the `bson` function.

Based on the time execution for the tests, it may be worth optimizing
the pattern matching code somewhat for which case comes first, avoiding
too many deeply nested active patterns. This is why the `|>` and `>>`
cases are defined separately.