Changed return type of update parser to option.
Rather than using `failwith` when dealing with an unrecognized
expression, simply returns `None`.

Also cleaned up implementation a bit. May eventually not want to use
the `updateParser` function with the computation expression patterns
simply because it is really more trouble it is worth to transform the
code quotations to that form.